\name{aperm.factor}
\alias{aperm.factor}
\title{An aperm method for factors}
\description{
hmmm...
}
\usage{
\method{aperm}{factor}(a, perm, ...)
}
\arguments{
  \item{a}{An array of mode factor.}
  \item{perm}{See \code{\link{aperm}}.}
  \item{\dots}{Additional arguments to pass to \code{\link{aperm}}.}
}
