\name{bci}
\alias{bci}
\docType{data}
\title{
Barro Colorado Island Data List
}
\description{
Abundances and functional traits of tree species on Barro Colorado Island.
}
\usage{data(bci)}
\source{
Traits are from:
Wright et. al (2010) Functional traits and the growth-mortality trade-off in tropical trees. Ecology. 91: 3664-3674.

Abundances are from:
Condit et. al (2002) Beta-diversity in tropical forest trees. Science 295: 666-669.
}
\note{
This data list object is mainly for illustration purposes and should probably not be used in serious scientific studies, because only species found in both the traits data set and the abundance data set are included.
}
\keyword{datasets}
