\name{rbind.rep}
\alias{cbind.rep}
\alias{rbind.rep}
\title{Matrix with identical rows or columns}
\usage{
  rbind.rep(x, times)

  cbind.rep(x, times)
}
\arguments{
  \item{x}{A vector}

  \item{times}{Number of times to replicate the vector}
}
\value{
  A matrix.
}
\description{
  Combining the \code{\link{rbind}} and \code{\link{cbind}}
  functions with the \code{\link{rep}} functions.
}

