\name{dimnames.data.list}
\alias{dimnames.data.list}
\alias{dimnames<-.data.list}
\title{Dimnames of a data list object}
\description{Retrieve or set the dimnames of a data list object.}
\usage{
\method{dimnames}{data.list}(x)
\method{dimnames}{data.list}(x) <- value
}
\arguments{
  \item{x}{A \code{\link{data.list}} object.}
  \item{value}{The dimnames for the benchmark variable in \code{x}.}
}
\seealso{
\code{\link{dimnames}} for the default method and \code{\link{dim.data.list}} to retrieve and set the dimensions of a data list.
}
