\name{higgins}
\alias{higgins}
\docType{data}
\title{
Stream fish
}
\description{
Real community data.
}
\usage{data(higgins)}

\source{
C. L. Higgins (2009) Spatiotemporal variation in functional and taxonomic organization of stream-fish assemblages in central Texas.  Aquatic Ecology. 43: 1133-1141.
}
\keyword{datasets}
