% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi-table.R
\name{multi_table}
\alias{multi_table}
\title{Write a list of tables as a single csv}
\usage{
multi_table(dat, file_path)
}
\arguments{
\item{file_path}{}
}
\description{
Write a list of tables as a single csv
}
