\name{multitable-package}
\alias{multitable-package}
\alias{multitable}
\docType{package}
\title{Simultaneous manipulation of multiple arrays of data, with data.list objects}
\description{
This package extends the concept of a data frame to more general objects called data lists.  While data frames are lists of data vectors, data lists are lists of data arrays (i.e. vectors with dim attributes).  In data frames the number of...
}
\details{
\tabular{ll}{
Package: \tab data.list\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-17\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
data.list is the most important function
}
\author{Steve Walker
Maintainer: Steve Walker <steve.walker@utoronto.ca>
}
\keyword{ package }
