\name{reorder2}
\alias{reorder2}
\title{More intuitive version of reorder}
\usage{
  reorder2(x, X)
}
\arguments{
  \item{x}{A factor.}

  \item{X}{Character vector of levels in the correct order.}
}
\value{
  Factor in the correct order.
}
\description{
  TODO:  actually i don't think this function is useful anymore.  there
  are some much better approaches in ggplot2 examples that use
  the factor function only
}

