\name{dim.data.list}
\alias{dim.data.list}
\alias{repdim}
\alias{dim<-.data.list}
\title{Extracting and assigning replication dimension information from data list objects}
\description{
The title says it all.
}
\usage{
repdim(x)

\method{dim}{data.list}(x)

\method{dim}{data.list}(x) <- value
}
\arguments{
  \item{x}{A data list.}
  \item{value}{An value for \code{value} will result in an error.}
}