\name{dimnames.data.list}
\alias{dimnames.data.list}
\alias{dimnames<-.data.list}
\title{Setting extracting dimnames in data lists}
\description{You get it.}
\usage{
\method{dimnames}{data.list}(x)
\method{dimnames}{data.list}(x) <- value
}
\arguments{
  \item{x}{A \code{\link{data.list}} object.}
  \item{value}{The dimnames for the benchmark variable in \code{x}.}
}
