\name{varnames}
\alias{varnames}
\alias{varnames<-}
\title{Set and extract variable names in data lists.}
\description{
You get it.
}
\usage{
varnames(dl)
varnames(dl) <- value
}
\arguments{
  \item{dl}{A \code{\link{data.list}} object.}
  \item{value}{A character vector with the names of the data list.}
}
